rm tree stack cache (too much memory overhead)
