+++ filetree/tree.go
@@ -3,6 +3,7 @@ package filetree
 import (
 	"fmt"
 	"github.com/google/uuid"
+	"github.com/sirupsen/logrus"
 	"sort"
 	"strings"
 )
@@ -298,3 +299,17 @@ func (tree *FileTree) markRemoved(path string) error {
 	}
 	return node.AssignDiffType(Removed)
 }
+
+// StackTreeRange combines an array of trees into a single tree
+func StackTreeRange(trees []*FileTree, start, stop int) *FileTree {
+
+	tree := trees[0].Copy()
+	for idx := start; idx <= stop; idx++ {
+		err := tree.Stack(trees[idx])
+		if err != nil {
+			logrus.Debug("could not stack tree range:", err)
+		}
+	}
+	return tree
+}
+
